// frontend/src/lib/api/generated/model/clientRead.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Datum API
 * OpenAPI spec version: 0.1.0
 */

export interface ClientRead {
  /**
     * @minLength 1
     * @maxLength 200
     */
  name: string;
  contact_person?: string | null;
  email?: string | null;
  phone?: string | null;
  address?: string | null;
  notes?: string | null;
  id: string;
  created_at: string;
  updated_at: string;
}